+++ examples/service/verify/bucket/create/bucket_create_example.go
@@ -1,0 +1,42 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v2 "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limit/buckets"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var verifySession *v2.Verify
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	verifySession = twilio.NewWithCredentials(creds).Verify.V2
+}
+
+func main() {
+	resp, err := verifySession.
+		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		RateLimit("RKXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		Buckets.
+		Create(&buckets.CreateBucketInput{
+			Max:      4,
+			Interval: 10,
+		})
+
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	log.Printf("SID: %s", resp.Sid)
+}