+++ service/verify/v2/service/rate_limit/api_op_client.go
@@ -1,7 +1,11 @@
 // Package rate_limit contains auto-generated files. DO NOT MODIFY
 package rate_limit
 
-import "github.com/RJPearson94/twilio-sdk-go/client"
+import (
+	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limit/bucket"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limit/buckets"
+)
 
 // Client for managing a specific rate limit resource
 // See https://www.twilio.com/docs/verify/api/service-rate-limits for more details
@@ -10,6 +14,9 @@ type Client struct {
 
 	serviceSid string
 	sid        string
+
+	Bucket  func(string) *bucket.Client
+	Buckets *buckets.Client
 }
 
 // ClientProperties are the properties required to manage the rate limit resources
@@ -25,5 +32,17 @@ func New(client *client.Client, properties ClientProperties) *Client {
 
 		serviceSid: properties.ServiceSid,
 		sid:        properties.Sid,
+
+		Bucket: func(bucketSid string) *bucket.Client {
+			return bucket.New(client, bucket.ClientProperties{
+				RateLimitSid: properties.Sid,
+				ServiceSid:   properties.ServiceSid,
+				Sid:          bucketSid,
+			})
+		},
+		Buckets: buckets.New(client, buckets.ClientProperties{
+			RateLimitSid: properties.Sid,
+			ServiceSid:   properties.ServiceSid,
+		}),
 	}
 }